Add contribution for integration with gitolite
Add in contrib/gitolite/patches-g2/ a patch to apply
to gitolite v2.x source such that gitolite can delegate
the read/update of some repositories to git-mirror proxy.

Add support for local-base configuration that can be used to
specify the prefix used for local repositories. This prefix will
be removed if the proxy forwards the request to the master server.