add support for auto-update and auto-create
Added support for auto-update in configuration.
Defaults to false.
If true, the mirror will be updated from the master
repository by a git fetch origin before serving the
upload from the mirror repository.
This adds the connection latency to the master, though
it ensures that the client fetch is always up-to-date
w.r.t. the master repository.

Added support for auto-create in configuration.
Defaults to false.
If true nd auto-update is also true, a missing
repository will be setup as a mirror of the master
repository.
This must be used with care when multiple account
access the mirror and have different credentials on
the master, as the account running the mirror may
have more priviledge than the client users.
This is the same limitation as the non auto-create
mode, but with auto-create, there is no more control
on which repository is actually mirrored. Any accessed
repository will be setup as mirrored and thus accessible
through the mirror proxy by all users.

Change-Id: I1b732f1de29d77b87079ae5e0aff8c97b4997d19