Add pruning when updating mirror with auto-update
The update of the mirrored repository is done with the --prune
option of git fetch, otherwise the mirror may have obsolete
references when some master references are deleted.

Add output of git fetch --prune origin command in logs for
diagnostic, as this command does not report a user facing failure,
generate a warning in the git-mirror.log if debug is set.

Change-Id: I847af6e89e0ffdcfd15eed0f6a95ab5f168bf1c5